requirements.txt
prometheus-client
paho-mqtt
smbus2
